+++ docs/changelog.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Drop support for Python 3.8 (#820)
 - Flag invalid regexes in arguments to functions like
   `re.search` (#816)